+++ CHANGES.md
@@ -1,7 +1,9 @@
 # GENI Portal Release Notes
 
 # [Release 3.9](https://github.com/GENI-NSF/geni-portal/milestones/3.9)
-
+* Fix bug where log and map tabs wouldn't show up if table view 
+  was preffered view on the dashboard.
+  ([#1653](https://github.com/GENI-NSF/geni-portal/issues/1653))
 
 # [Release 3.8](https://github.com/GENI-NSF/geni-portal/milestones/3.8)