+++ CHANGES.md
@@ -4,6 +4,8 @@
 
 * Add "New Project" button to project table view.
   ([#1650](https://github.com/GENI-NSF/geni-portal/issues/1650))
+* Make project table view show expired projects.
+  ([#1651](https://github.com/GENI-NSF/geni-portal/issues/1651))
 
 # [Release 3.8](https://github.com/GENI-NSF/geni-portal/milestones/3.8)